requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
Pillow==8.3.2
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
django-tinymce==3.5.0
django-filter==23.2
django-storages==1.13.2
boto3==1.26.151
fontawesomefree==6.3.0
django-crispy-forms==2.0
six==1.16.0
gunicorn==20.1.0
whitenoise==6.4.0
psycopg2-binary==2.9.1
djangorestframework-simplejwt==5.2.2
crispy-bootstrap5==0.7
python-decouple
